refactor: split game package mode wrappers and types
